# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi-Tenant Express API",
    "description": "Multi-tenant REST API with Organization → Project → Resource hierarchy, RBAC permissions, and invitation system.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "description": "Rotate tokens — revokes the current refresh token cookie and sets new httpOnly cookies.",
        "operationId": "refreshToken",
        "security": [
          {
            "RefreshToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token refreshed successfully. New auth cookies are set.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              },
              "Set-Cookie": {
                "description": "Rotated httpOnly auth cookies (access_token and refresh_token)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "message": "OK",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "JWT access token (httpOnly cookie set by server)"
      },
      "RefreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "JWT refresh token (httpOnly cookie set by server)"
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Correlation ID for request tracing. Accepts a valid UUID or generates one.",
        "schema": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": ["message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
...
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or invalid credentials",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "noToken": {
                "summary": "No token provided",
                "value": {
                  "message": "No token provided"
                }
              },
              "invalid": {
                "summary": "Invalid token",
                "value": {
                  "message": "Invalid token"
                }
              },
              "expired": {
                "summary": "Token expired",
                "value": {
                  "message": "Token expired"
                }
              },
              "credentials": {
                "summary": "Invalid credentials",
                "value": {
                  "message": "invalid credentials"
                }
              }
            }
          }
        }
      },
...
    }
  }
}